+++ pkg/analysis_server/test/integration/support/integration_tests.dart
@@ -712,6 +712,7 @@ class Server {
     if (Platform.packageConfig != null) {
       arguments.add('--packages=${Platform.packageConfig}');
     }
+    arguments.add('--disable-service-auth-codes');
     //
     // Add the server executable.
     //